Fix is_same_struct check.
First, added a fast path: If the type pointers are equal, the types are
obivously the same.

Second, the code used to treat the vectors returned by dict_keys as vectors
of (Type *), but it is clear that they are vectors of (char *), since that
is the only type dict_put can put into the vector of keys. This type
confusion was not caught by the compiler due to the implicit conversions to
and from (void *) for vector elements. To obtain actual (Type *) pointers,
another indirection through dict_get is needed.